Splitted spring beans configuration
